requirements-test.txt
coverage>=5
mock>=4.0.3
flake8~=3.5
testfixtures>=6.18.5
pytest~=6.2.5
pytest-cov~=3.0.0
pytest-mock~=3.6.1
pytest-xdist>=1.34.0
pytest-rerunfailures~=10.2
pytest-timeout~=2.0.2
pytest-benchmark~=3.4.0
easydict>=1.7,<2
testtools>=2
where>=1.0.2
pytest-image-diff>=0.0.11
matplotlib
natsort
git+https://github.com/narugo1992/responses.git@dev/binary
git+https://github.com/narugo1992/pytest_httpx_recorder.git@main
click>=8.0.0
gchar>=0.1.0
hfutils>=0.0.4
pytest_httpx